Added toggling functionality to Control Intake
Added the ability to toggle the actuator on a button press in Control
Intake, and created two different directions of spinning the intake
depending on which button is pressed.
Moved the implementation of ToggleClass to the .h file so that the
compiler is able to access the definitions of the various forms of the
template from the inclusion of the .h file in the files making use of
it.