Modified Climber, Added POV to RobotMap, and Updated Button Settings
Modified Climber to run off of the POV that was added to RobotMap, and
to have two separate motors that run to control the winch and tapes
separately. Also updated the Button Maps to fit the current requests by
the Drivers. Switched all Commands away from using the LST variables,
except for the POV, due to its functionality.