Reconfiguration of autonPosition Chooser
Changed the return values of autonPosition to be the integer value of
the selected defense. To receive angle created another function within
OI that uses a switch to return the angle depending on the selected
auton. Made intake vertical position change depending on which defense
position is being ran, with it only being down for low bar.