Added commands to put the intake into specific positions
Added two commands that place the intake into a the CDF position and the
portcullis position as long as they are held. When they are released
they drive the intake fully in. To run commands off of an AXS it was
necessary to define a class derived from Trigger, which was done within
OI.h as it is unlikely to be needed outside of OI.