Added a button for testing driving to point in both gears.
Made button 10 on the left joystick a testing button similar to button
10 on the right one, in that it runs a test command useful for tuning
PID by running the system making use of the PID in a consistent manner.
This button drives to robot forward 20 inches, making use of the
AutonDriveToPoint command. To make it run in whatever gear is currently
being used, a getter for the state of the gear shifting was added to
Chassis.h. Also added an option to AutonDriveToPoint in the constructor,
defaulted to the method that it used to run, allowing it to not set the
shift to any position, rather leaving it as it was in. As a minor change
made in this cycle, CameraAim was made to not shift, and just stay in
whatever gear the robot is currently in.